career_agent/requirements.txt
gradio
python-dotenv
langchain
langchain-google-genai
langchain-community
google-generativeai